Proofreading and commentary
To keep in line with the convention established for this edition, I
added details of the proofreading and commentary inline.

Most of the changes are to resolve differences between the sources, but
I also made a couple corrections.

For a few spots where later editions differ from listed sources, I added
inline notes and mention to refer to mUltimate Chopin to view additional
commentary about the differences.